pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "black>=26.3.1",
    "fastapi>=0.135.3",
    "flake8>=7.3.0",
    "isort>=8.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.44.0",
]
